+++ CHANGES.rst
@@ -5,6 +5,9 @@
 6.4.1 (unreleased)
 ==================
 
+- ``unittest.TestCase.subTest`` support
+  (`#91 <https://github.com/zopefoundation/zope.testrunner/issues/91>`_).
+
 
 6.4 (2024-02-27)
 ================